Fixed serving from my ARM64 Debian 12.10 VM.
It was an issue with `jq` syntax for indexing lists in the justfile.
I'm sure this worked at one point, and I can only assume it was because
I was running a different version of `jq`. A comment in the justfile
reminded me I was at one point running this on Ubuntu, so maybe Ubuntu's
version of jq supported that syntax? But that's *just* a guess.

Anyway, I significantly prefer this syntax :). Hope it continues to
work!